Fixed bugs for tutor's courses, subjects & area mapping
Fixed some bugs for tutors 'courses, subjects and areas cover mapping that include the new system table:-
        1) tk_tution_course
	2) tk_tution_subject
	3) tk_tutor_area_cover
	4) tk_tutor_request
	5) tk_tutor_subject
The bugs of those codes are the code didn't import the data, missed match of mapping, and not well organized structure. The fixed also keep watching the rows to avoid the redundancy and clone information import by the script. INSERT IGNORE INTO query added to double check and skip the information that already imported. This could save the time of executing the script.